Extraceted the comment widget out
